Add some Database Fetures

+++ app/src/main/java/com/hevs/gym/fitnessapp/MyGroupsActivity.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 
 public class MyGroupsActivity extends AppCompatActivity {
 
-    private int idUser;
+    private long idUser;
     private ArrayList<Button> buttonList;
 
     @Override
